fix(nextjs): ignore node i18n generated functions
