+++ lib/presets/custom/next/compute/default/prebuild/mapping/index.js
@@ -7,7 +7,7 @@ import { feedback } from '#utils';
 import { validateFile, normalizePath } from '../../../utils/fs.js';
 import { formatRoutePath, stripIndexRoute } from '../../../utils/routing.js';
 import { fixFunctionContent } from '../edge/index.js';
-import getNextProjectConfig from '../../../utils/next.js';
+import { getNextProjectConfig, isLocalePath } from '../../../utils/next.js';
 
 /**
  *
@@ -105,21 +105,6 @@ async function mapAndAdaptFunction(
   const functionsDir = join('.vercel', 'output', 'functions');
   const path = vcObject.path.replace('/.vc-config.json', '');
 
-  const nextProjectConfig = await getNextProjectConfig();
-
-  const locales = nextProjectConfig?.i18n?.locales;
-  if (locales && locales?.length > 0) {
-    let isLangRoute = false;
-    locales.forEach((lang) => {
-      if (path.includes(lang)) {
-        isLangRoute = true;
-      }
-    });
-    if (isLangRoute) {
-      return {};
-    }
-  }
-
   const { runtime } = vcObject.content;
   const isEdge = runtime.match(/edge/);
   const isNode = runtime.match(/node/);
@@ -205,12 +190,19 @@ export async function mapAndAdaptFunctions(
   const vcConfigPaths = glob.sync(
     '.vercel/output/functions/**/.vc-config.json',
   );
-  const vcConfigObjects = vcConfigPaths.map((file) => {
+  let vcConfigObjects = vcConfigPaths.map((file) => {
     return {
       path: file,
       content: JSON.parse(readFileSync(file, 'utf8')),
     };
   });
+
+  // filter i18n aditional functions
+  const nextProjectConfig = await getNextProjectConfig();
+  const locales = nextProjectConfig?.i18n?.locales || [];
+  vcConfigObjects = vcConfigObjects.filter(
+    (config) => !isLocalePath(config.path, locales),
+  );
 
   const vcObjects = {
     invalid: vcConfigObjects.filter(
