+++ lib/presets/custom/next/compute/prebuild.js
@@ -12,7 +12,7 @@ import {
 
 import runDefaultBuild from './default/prebuild/index.js';
 import runNodeBuild from './node/prebuild/index.js';
-import getNextProjectConfig from './utils/next.js';
+import { getNextProjectConfig, isLocalePath } from './utils/next.js';
 
 const { deleteTelemetryFiles, createVercelProjectConfig, runVercelBuild } =
   VercelUtils;
@@ -36,26 +36,6 @@ async function vercelPrebuildActions() {
     feedback.prebuild.error(error);
     process.exit(1);
   }
-}
-
-/**
- * Detect if a path is a locale path based on locales array
- * @param {string} path - path to be checked
- * @param {Array} locales - locales path
- * @returns {boolean} - indicates if is a locale path
- */
-function isLocalePath(path, locales) {
-  let isLangRoute = false;
-
-  if (locales && locales?.length > 0) {
-    locales.forEach((lang) => {
-      if (path.includes(lang)) {
-        isLangRoute = true;
-      }
-    });
-  }
-
-  return isLangRoute;
 }
 
 /**
