+++ lib/presets/custom/next/compute/utils/next.js
@@ -18,4 +18,24 @@ async function getNextProjectConfig() {
   return config;
 }
 
-export default getNextProjectConfig;
+/**
+ * Detect if a path is a locale path based on locales array
+ * @param {string} path - path to be checked
+ * @param {Array} locales - locales path
+ * @returns {boolean} - indicates if is a locale path
+ */
+function isLocalePath(path, locales) {
+  let isLangRoute = false;
+
+  if (locales && locales?.length > 0) {
+    locales.forEach((lang) => {
+      if (path.includes(lang)) {
+        isLangRoute = true;
+      }
+    });
+  }
+
+  return isLangRoute;
+}
+
+export { getNextProjectConfig, isLocalePath };
